Move RunReport into its own module
