+++ tests/monitoring/test_monitor_loop.py
@@ -15,11 +15,11 @@ from unittest.mock import MagicMock, Mock, patch
 
 import pytest
 
-from gcmon.data import RunReport
 from gcmon.monitor import PollReport
 from gcmon.monitor_loop import MonitorLoop
 from gcmon.rss_sampler import RssSampler
 from gcmon.run_policy import InfinityRunner, Runner
+from gcmon.run_report import RunReport
 
 
 def _report(*live: int, keep_running: bool = True) -> PollReport:
